+++ release-notes.md
@@ -1,4 +1,5 @@
 Changes since v1.3.2 include
+- lsusb output box has more verbose info for easier debugging
 - make 'other' checkbox for SDRs stick when selected
 - fix routeAPI not working
 - fix FR24 sometimes being displayed in aggregator links when not enabled